requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
pyyaml==6.0.1
coreapi==2.3.3
drf-yasg==1.21.7
